feat: webclient 요청 시 타임아웃 시간 설정 기능 작성

+++ backend/src/main/java/com/shyashyashya/refit/domain/qnaset/constant/StarAnalysisConstant.java
@@ -4,5 +4,5 @@ public final class StarAnalysisConstant {
 
     private StarAnalysisConstant() {}
 
-    public static final int STAR_ANALYSIS_CREATE_REQUEST_TIMEOUT_SEC = 30;
+    public static final long STAR_ANALYSIS_CREATE_REQUEST_TIMEOUT_SEC = 30;
 }
